+++ packages/yume-engine/src/bundle/index.ts
@@ -2,3 +2,4 @@ export * from './xml-plist-parser.js';
 export * from './bplist-parser.js';
 export * from './plist-parser.js';
 export * from './app-bundle-inspector.js';
+export * from './bundle-classifier.js';